Switched to fetching double directly
Build the finish function according to the cpp implementation

Removed commented out code
